Refactor AnalyzeNs()
