+++ src/php/Analyzer/AnalyzeTuple.php
@@ -7,11 +7,11 @@ namespace Phel\Analyzer;
 use Exception;
 use Phel\Analyzer;
 use Phel\Analyzer\AnalyzeTuple\AnalyzeDef;
+use Phel\Analyzer\AnalyzeTuple\AnalyzeNs;
 use Phel\Ast\ApplyNode;
 use Phel\Ast\BindingNode;
 use Phel\Ast\CallNode;
 use Phel\Ast\CatchNode;
-use Phel\Ast\DefNode;
 use Phel\Ast\DefStructNode;
 use Phel\Ast\DoNode;
 use Phel\Ast\FnNode;
@@ -21,7 +21,6 @@ use Phel\Ast\IfNode;
 use Phel\Ast\LetNode;
 use Phel\Ast\MethodCallNode;
 use Phel\Ast\Node;
-use Phel\Ast\NsNode;
 use Phel\Ast\PhelArrayNode;
 use Phel\Ast\PhpArrayGetNode;
 use Phel\Ast\PhpArrayPushNode;
@@ -36,11 +35,9 @@ use Phel\Ast\ThrowNode;
 use Phel\Ast\TryNode;
 use Phel\Destructure;
 use Phel\Exceptions\AnalyzerException;
-use Phel\Lang\Keyword;
 use Phel\Lang\AbstractType;
 use Phel\Lang\PhelArray;
 use Phel\Lang\Symbol;
-use Phel\Lang\Table;
 use Phel\Lang\Tuple;
 use Phel\NodeEnvironment;
 use Phel\RecurFrame;
@@ -64,7 +61,7 @@ final class AnalyzeTuple
             case 'def':
                 return (new AnalyzeDef($this->analyzer))($x, $env);
             case 'ns':
-                return $this->analyzeNs($x, $env);
+                return (new AnalyzeNs($this->analyzer))($x, $env);
             case 'fn':
                 return $this->analyzeFn($x, $env);
             case 'quote':
@@ -204,120 +201,6 @@ final class AnalyzeTuple
                 $x->setEndLocation($parent->getEndLocation());
             }
         }
-    }
-
-    private function analyzeNs(Tuple $x, NodeEnvironment $env): NsNode
-    {
-        $tupleCount = count($x);
-        if (!($x[1] instanceof Symbol)) {
-            throw new AnalyzerException(
-                "First argument of 'ns must be a Symbol",
-                $x->getStartLocation(),
-                $x->getEndLocation()
-            );
-        }
-
-        $ns = $x[1]->getName();
-        if (!(preg_match("/^[a-zA-Z\x7f-\xff][a-zA-Z0-9\-\x7f-\xff\\\\]*[a-zA-Z0-9\-\x7f-\xff]*$/", $ns))) {
-            throw new AnalyzerException(
-                'The namespace is not valid. A valid namespace name starts with a letter,
-                followed by any number of letters, numbers, or dashes. Elements are splitted by a backslash.',
-                $x[1]->getStartLocation(),
-                $x[1]->getEndLocation()
-            );
-        }
-
-        $parts = explode('\\', $ns);
-        foreach ($parts as $part) {
-            if ($this->isPHPKeyword($part)) {
-                throw new AnalyzerException(
-                    "The namespace is not valid. The part '$part' can not be used because it is a reserved keyword.",
-                    $x[1]->getStartLocation(),
-                    $x[1]->getEndLocation()
-                );
-            }
-        }
-
-        $this->analyzer->getGlobalEnvironment()->setNs($ns);
-
-        $requireNs = [];
-        for ($i = 2; $i < $tupleCount; $i++) {
-            $import = $x[$i];
-
-            if (!($import instanceof Tuple)) {
-                throw new AnalyzerException(
-                    "Import in 'ns must be Tuples.",
-                    $x->getStartLocation(),
-                    $x->getEndLocation()
-                );
-            }
-
-            /** @var Tuple $import */
-            if ($this->isKeywordWithName($import[0], 'use')) {
-                if (!($import[1] instanceof Symbol)) {
-                    throw new AnalyzerException(
-                        'First arugment in :use must be a symbol.',
-                        $import->getStartLocation(),
-                        $import->getEndLocation()
-                    );
-                }
-
-                if (count($import) === 4 && $this->isKeywordWithName($import[2], 'as')) {
-                    $alias = $import[3];
-                    if (!($alias instanceof Symbol)) {
-                        throw new AnalyzerException(
-                            'Alias must be a Symbol',
-                            $import->getStartLocation(),
-                            $import->getEndLocation()
-                        );
-                    }
-                } else {
-                    $parts = explode('\\', $import[1]->getName());
-                    $alias = new Symbol($parts[count($parts) - 1]);
-                }
-
-                $this->analyzer->getGlobalEnvironment()->addUseAlias($ns, $alias, $import[1]);
-            } elseif ($this->isKeywordWithName($import[0], 'require')) {
-                if (!($import[1] instanceof Symbol)) {
-                    throw new AnalyzerException(
-                        'First arugment in :require must be a symbol.',
-                        $import->getStartLocation(),
-                        $import->getEndLocation()
-                    );
-                }
-
-                $requireNs[] = $import[1];
-
-                if (count($import) === 4 && $this->isKeywordWithName($import[2], 'as')) {
-                    $alias = $import[3];
-                    if (!($alias instanceof Symbol)) {
-                        throw new AnalyzerException(
-                            'Alias must be a Symbol',
-                            $import->getStartLocation(),
-                            $import->getEndLocation()
-                        );
-                    }
-                } else {
-                    $parts = explode('\\', $import[1]->getName());
-                    $alias = new Symbol($parts[count($parts) - 1]);
-                }
-
-                $this->analyzer->getGlobalEnvironment()->addRequireAlias($ns, $alias, $import[1]);
-            }
-        }
-
-        return new NsNode($x[1]->getName(), $requireNs, $x->getStartLocation());
-    }
-
-    private function isPHPKeyword(string $w): bool
-    {
-        return in_array($w, PhpKeywords::KEYWORDS, true);
-    }
-
-    /** @param mixed $x */
-    private function isKeywordWithName($x, string $name): bool
-    {
-        return $x instanceof Keyword && $x->getName() === $name;
     }
 
     private function analyzeFn(Tuple $x, NodeEnvironment $env): FnNode
